Feat : 보스 체력바 연동

+++ Assets/2.Private/KimSW/Scripts/InGameUI.cs
@@ -209,7 +209,7 @@ public class InGameUI : BindUI
     /// 적 체력 초기화
     /// </summary>
     /// <param name="hp"></param>
-    public void InitEnemyHP(int hp)
+    public void InitEnemyHP(float hp)
     {
         enemyStatusPanel.SetEnemyHP(hp);
         OnEnemyHP();
